src/requirements.txt
binaryornot==0.4.4
cachetools==5.3.0
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
coloredlogs==15.0.1
contourpy==1.0.7
cycler==0.11.0
deepdiff==6.2.2
docker==6.1.2
fonttools==4.39.4
google-auth==2.18.0
humanfriendly==10.0
idna==3.4
kathara==3.5.5
kiwisolver==1.4.4
kubernetes==26.1.0
matplotlib==3.7.1
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
ordered-set==4.1.0
packaging==23.1
Pillow==9.5.0
progressbar2==4.2.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
pyroute2==0.7.8
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-utils==3.5.2
PyYAML==6.0
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
slug==2.0
terminaltables==3.1.10
urllib3==1.26.15
websocket-client==1.5.1
